Made TradeOffer class exportable
