+++ lib/classes/TradeOffer.js
@@ -5,184 +5,10 @@ var SteamID = require('steamid');
 var EconItem = require('./EconItem.js');
 var Helpers = require('../helpers.js');
 
-var ETradeOfferState = TradeOfferManager.ETradeOfferState;
-var EOfferFilter = TradeOfferManager.EOfferFilter;
-var EConfirmationMethod = TradeOfferManager.EConfirmationMethod;
-var ETradeStatus = TradeOfferManager.ETradeStatus;
-
-TradeOfferManager.prototype.createOffer = function(partner, token) {
-	if (typeof partner === 'string' && partner.match(/^https?:\/\//)) {
-		// It's a trade URL I guess
-		var url = require('url').parse(partner, true);
-		if (!url.query.partner) {
-			throw new Error("Invalid trade URL");
-		}
-
-		partner = SteamID.fromIndividualAccountID(url.query.partner);
-		token = url.query.token;
-	}
-
-	var offer = new TradeOffer(this, partner, token);
-	offer.isOurOffer = true;
-	offer.fromRealTimeTrade = false;
-	return offer;
-};
-
-TradeOfferManager.prototype.getOffer = function(id, callback) {
-	this._apiCall('GET', 'GetTradeOffer', 1, {"tradeofferid": id}, (err, body) => {
-		if (err) {
-			callback(err);
-			return;
-		}
-
-		if (!body.response) {
-			callback(new Error("Malformed API response"));
-			return;
-		}
-
-		if (!body.response.offer) {
-			callback(new Error("No matching offer found"));
-			return;
-		}
-
-		// Make sure the response is well-formed
-		if (offerMalformed(body.response.offer)) {
-			callback(new Error("Data temporarily unavailable"));
-			return;
-		}
-
-		this._digestDescriptions(body.response.descriptions);
-		checkNeededDescriptions(this, [body.response.offer], (err) => {
-			if (err) {
-				callback(err);
-				return;
-			}
-
-			callback(null, createOfferFromData(this, body.response.offer));
-		});
-	});
-};
-
-TradeOfferManager.prototype.getOffers = function(filter, historicalCutoff, callback) {
-	if (typeof historicalCutoff === 'function') {
-		callback = historicalCutoff;
-		historicalCutoff = new Date(Date.now() + 31536000000);
-	} else if (!historicalCutoff) {
-		historicalCutoff = new Date(Date.now() + 31536000000);
-	}
-
-	// Currently the GetTradeOffers API doesn't include app_data, so we need to get descriptions from the WebAPI
-
-	var options = {
-		"get_sent_offers": 1,
-		"get_received_offers": 1,
-		"get_descriptions": 0,
-		"language": this._language,
-		"active_only": (filter == EOfferFilter.ActiveOnly) ? 1 : 0,
-		"historical_only": (filter == EOfferFilter.HistoricalOnly) ? 1 : 0,
-		"time_historical_cutoff": Math.floor(historicalCutoff.getTime() / 1000)
-	};
-
-	this._apiCall('GET', 'GetTradeOffers', 1, options, (err, body) => {
-		if (err) {
-			callback(err);
-			return;
-		}
-
-		if (!body.response) {
-			callback(new Error("Malformed API response"));
-			return;
-		}
-
-		// Make sure at least some offers are well-formed. Apparently some offers can be empty just forever. Because Steam.
-		var allOffers = (body.response.trade_offers_sent || []).concat(body.response.trade_offers_received || []);
-		if (allOffers.length > 0 && (allOffers.every(offerMalformed) || allOffers.some(offerSuperMalformed))) {
-			callback(new Error("Data temporarily unavailable"));
-			return;
-		}
-
-		//manager._digestDescriptions(body.response.descriptions);
-
-		// Let's check the asset cache and see if we have descriptions that match these items.
-		// If the necessary descriptions aren't in the asset cache, this will request them from the WebAPI and store
-		// them for future use.
-		checkNeededDescriptions(this, (body.response.trade_offers_sent || []).concat(body.response.trade_offers_received || []), (err) => {
-			if (err) {
-				callback(new Error("Descriptions: " + err.message));
-				return;
-			}
-
-			var sent = (body.response.trade_offers_sent || []).map(data => createOfferFromData(this, data));
-			var received = (body.response.trade_offers_received || []).map(data => createOfferFromData(this, data));
-
-			callback(null, sent, received);
-			this.emit('offerList', filter, sent, received);
-		});
-	});
-};
-
-function offerSuperMalformed(offer) {
-	return !offer.accountid_other;
-}
-
-function offerMalformed(offer) {
-	return offerSuperMalformed(offer) || ((offer.items_to_give || []).length == 0 && (offer.items_to_receive || []).length == 0);
-}
-
-function createOfferFromData(manager, data) {
-	var offer = new TradeOffer(manager, new SteamID('[U:1:' + data.accountid_other + ']'));
-	offer.id = data.tradeofferid.toString();
-	offer.message = data.message;
-	offer.state = data.trade_offer_state;
-	offer.itemsToGive = data.items_to_give || [];
-	offer.itemsToReceive = data.items_to_receive || [];
-	offer.isOurOffer = data.is_our_offer;
-	offer.created = new Date(data.time_created * 1000);
-	offer.updated = new Date(data.time_updated * 1000);
-	offer.expires = new Date(data.expiration_time * 1000);
-	offer.tradeID = data.tradeid ? data.tradeid.toString() : null;
-	offer.fromRealTimeTrade = data.from_real_time_trade;
-	offer.confirmationMethod = data.confirmation_method || EConfirmationMethod.None;
-	offer.escrowEnds = data.escrow_end_date ? new Date(data.escrow_end_date * 1000) : null;
-	offer.rawJson = JSON.stringify(data, null, "\t");
-
-	if (manager._language) {
-		offer.itemsToGive = manager._mapItemsToDescriptions(null, null, offer.itemsToGive);
-		offer.itemsToReceive = manager._mapItemsToDescriptions(null, null, offer.itemsToReceive);
-	} else {
-		offer.itemsToGive = processItems(offer.itemsToGive);
-		offer.itemsToReceive = processItems(offer.itemsToReceive);
-	}
-
-	return offer;
-}
-
-function processItems(items) {
-	return items.map(item => new EconItem(item));
-}
-
-function checkNeededDescriptions(manager, offers, callback) {
-	if (!manager._language) {
-		callback(null);
-		return;
-	}
-
-	var items = [];
-	offers.forEach((offer) => {
-		(offer.items_to_give || []).concat(offer.items_to_receive || []).forEach((item) => {
-			if (!manager._hasDescription(item)) {
-				items.push(item);
-			}
-		});
-	});
-
-	if (!items.length) {
-		callback(null);
-		return;
-	}
-
-	manager._requestDescriptions(items, callback);
-}
+const ETradeOfferState = require('../../resources/ETradeOfferState.js');
+const EOfferFilter = require('../../resources/EOfferFilter.js');
+const EConfirmationMethod = require('../../resources/EConfirmationMethod.js');
+const ETradeStatus = require('../../resources/ETradeStatus.js');
 
 function TradeOffer(manager, partner, token) {
 	if (typeof partner === 'string') {
@@ -804,11 +630,11 @@ TradeOffer.prototype.getReceivedItems = function(getActions, callback) {
 		}
 
 		if (!getActions) {
-			callback(null, processItems(items));
+			callback(null, Helpers.processItems(items));
 		} else {
 			this.manager._addDescriptions(items, (err, describedItems) => {
 				if (err) {
-					callback(null, processItems(items)); // welp, we have to just accept what we have
+					callback(null, Helpers.processItems(items)); // welp, we have to just accept what we have
 				} else {
 					callback(null, describedItems);
 				}
@@ -1014,3 +840,5 @@ TradeOffer.prototype.setToken = function(token) {
 
 	this._token = token;
 };
+
+module.exports = TradeOffer;
