+++ lib/index.js
@@ -12,6 +12,8 @@ const EResult = TradeOfferManager.EResult = require('../resources/EResult.js');
 const EConfirmationMethod = TradeOfferManager.EConfirmationMethod = require('../resources/EConfirmationMethod.js');
 const ETradeStatus = TradeOfferManager.ETradeStatus = require('../resources/ETradeStatus.js');
 
+const TradeOffer = require('./classes/TradeOffer.js');
+
 require('util').inherits(TradeOfferManager, require('events').EventEmitter);
 
 function TradeOfferManager(options) {
@@ -245,7 +247,117 @@ TradeOfferManager.prototype.getOffersContainingItems = function(items, includeIn
 	}
 };
 
+TradeOfferManager.prototype.createOffer = function(partner, token) {
+	if (typeof partner === 'string' && partner.match(/^https?:\/\//)) {
+		// It's a trade URL I guess
+		var url = require('url').parse(partner, true);
+		if (!url.query.partner) {
+			throw new Error("Invalid trade URL");
+		}
+
+		partner = SteamID.fromIndividualAccountID(url.query.partner);
+		token = url.query.token;
+	}
+
+	var offer = new TradeOffer(this, partner, token);
+	offer.isOurOffer = true;
+	offer.fromRealTimeTrade = false;
+	return offer;
+};
+
 require('./webapi.js');
 require('./assets.js');
 require('./polling.js');
-require('./classes/TradeOffer.js');
+
+TradeOfferManager.prototype.getOffer = function(id, callback) {
+	this._apiCall('GET', 'GetTradeOffer', 1, {"tradeofferid": id}, (err, body) => {
+		if (err) {
+			callback(err);
+			return;
+		}
+
+		if (!body.response) {
+			callback(new Error("Malformed API response"));
+			return;
+		}
+
+		if (!body.response.offer) {
+			callback(new Error("No matching offer found"));
+			return;
+		}
+
+		// Make sure the response is well-formed
+		if (Helpers.offerMalformed(body.response.offer)) {
+			callback(new Error("Data temporarily unavailable"));
+			return;
+		}
+
+		this._digestDescriptions(body.response.descriptions);
+		Helpers.checkNeededDescriptions(this, [body.response.offer], (err) => {
+			if (err) {
+				callback(err);
+				return;
+			}
+
+			callback(null, Helpers.createOfferFromData(this, body.response.offer));
+		});
+	});
+};
+
+TradeOfferManager.prototype.getOffers = function(filter, historicalCutoff, callback) {
+	if (typeof historicalCutoff === 'function') {
+		callback = historicalCutoff;
+		historicalCutoff = new Date(Date.now() + 31536000000);
+	} else if (!historicalCutoff) {
+		historicalCutoff = new Date(Date.now() + 31536000000);
+	}
+
+	// Currently the GetTradeOffers API doesn't include app_data, so we need to get descriptions from the WebAPI
+
+	var options = {
+		"get_sent_offers": 1,
+		"get_received_offers": 1,
+		"get_descriptions": 0,
+		"language": this._language,
+		"active_only": (filter == EOfferFilter.ActiveOnly) ? 1 : 0,
+		"historical_only": (filter == EOfferFilter.HistoricalOnly) ? 1 : 0,
+		"time_historical_cutoff": Math.floor(historicalCutoff.getTime() / 1000)
+	};
+
+	this._apiCall('GET', 'GetTradeOffers', 1, options, (err, body) => {
+		if (err) {
+			callback(err);
+			return;
+		}
+
+		if (!body.response) {
+			callback(new Error("Malformed API response"));
+			return;
+		}
+
+		// Make sure at least some offers are well-formed. Apparently some offers can be empty just forever. Because Steam.
+		var allOffers = (body.response.trade_offers_sent || []).concat(body.response.trade_offers_received || []);
+		if (allOffers.length > 0 && (allOffers.every(Helpers.offerMalformed) || allOffers.some(Helpers.offerSuperMalformed))) {
+			callback(new Error("Data temporarily unavailable"));
+			return;
+		}
+
+		//manager._digestDescriptions(body.response.descriptions);
+
+		// Let's check the asset cache and see if we have descriptions that match these items.
+		// If the necessary descriptions aren't in the asset cache, this will request them from the WebAPI and store
+		// them for future use.
+		Helpers.checkNeededDescriptions(this, (body.response.trade_offers_sent || []).concat(body.response.trade_offers_received || []), (err) => {
+			if (err) {
+				callback(new Error("Descriptions: " + err.message));
+				return;
+			}
+
+			var sent = (body.response.trade_offers_sent || []).map(data => Helpers.createOfferFromData(this, data));
+			var received = (body.response.trade_offers_received || []).map(data => Helpers.createOfferFromData(this, data));
+
+			callback(null, sent, received);
+			this.emit('offerList', filter, sent, received);
+		});
+	});
+};
